requirements.txt
aenum==3.1.15
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiorun==2024.8.1
aiosignal==1.3.2
async-timeout==5.0.1
attrs==24.3.0
coloredlogs==15.0.1
dacite==1.8.1
frozenlist==1.5.0
home-assistant-chip-clusters==2024.11.4
humanfriendly==10.0
idna==3.10
multidict==6.1.0
orjson==3.10.13
propcache==0.2.1
python-matter-server==7.0.1
yarl==1.18.3
